Planner buttons are working with the backend. Logic in place for adding only 4 and not deleting only 1.

+++ src/components/Planner/WeeklyPlannerBox/RemoveWeeklyPlannerButton.jsx
@@ -1,13 +1,17 @@
 import axios from "axios";
 
-const RemoveWeeklyPlannerButton = ({ weekId, setUserWeeks }) => {
+const RemoveWeeklyPlannerButton = ({ weekId, setUserWeeks, userWeeks }) => {
   const handleRemoveWeek = async () => {
     // make backend call to remove current week
-    const res = await axios.delete(`/api/delete-week/${weekId}`);
-    console.log(res.data);
+    if (userWeeks.length > 1) {
+      const res = await axios.delete(`/api/delete-week/${weekId}`);
+      console.log(res.data);
 
-    if (res.data.success) {
-      setUserWeeks(res.data.userWeeks);
+      if (res.data.success) {
+        setUserWeeks(res.data.userWeeks);
+      }
+    } else {
+      alert("Don't delete your only week!");
     }
   };
